Conflicts:
	scripts/fixtures/radar_fixtures/__init__.py

+++ radar/ukrdc_exporter/__main__.py
@@ -44,28 +44,34 @@ order by x.id, patients.id
 """.format(changed_sql)
 
 
-def export_patients(last_log_id=0):
-    if last_log_id == 0:
-        sql = all_sql
-    else:
-        sql = changed_sql
+def export_all(last_log_id):
+    return export_query(all_sql, last_log_id)
 
+
+def export_changed(last_log_id):
+    return export_query(changed_sql, last_log_id)
+
+
+def export_query(sql, last_log_id):
     rows = db.session.execute(sql, {'last_log_id': last_log_id})
 
-    patient_ids = set()
+    patient_ids = []
 
     for log_id, patient_id in rows:
         if log_id is not None:
             last_log_id = log_id
 
-        if patient_id in patient_ids:
-            continue
+        patient_ids.append(patient_id)
 
-        logger.info('Adding patient to queue id={}'.format(patient_id))
-        export_to_ukrdc(patient_id)
-        patient_ids.add(patient_id)
+    export_patients(patient_ids)
 
     return last_log_id
+
+
+def export_patients(patient_ids):
+    for patient_id in sorted(set(patient_ids)):
+        logger.info('Adding patient to queue id={}'.format(patient_id))
+        export_to_ukrdc(patient_id)
 
 
 def lock(f):
@@ -76,39 +82,44 @@ def main():
     logging.basicConfig(level=logging.INFO)
 
     parser = argparse.ArgumentParser()
-    parser.add_argument('--all', action='store_true')
+    group = parser.add_mutually_exclusive_group()
+    group.add_argument('--all', action='store_true')
+    group.add_argument('--changed', action='store_true')
+    group.add_argument('--id', type=int, dest='patient_ids', action='append', default=[])
     args = parser.parse_args()
 
     app = RadarUKRDCExporter()
 
-    state = app.config['UKRDC_EXPORTER_STATE']
+    if args.all or args.changed:
+        state = app.config['UKRDC_EXPORTER_STATE']
 
-    if state is not None:
-        if not os.path.exists(state):
-            with open(state, 'w') as state_f:
-                state_f.write(str(0))
-                state_f.write('\n')
+        if state is not None:
+            if not os.path.exists(state):
+                with open(state, 'w') as state_f:
+                    state_f.write(str(0))
+                    state_f.write('\n')
 
-        state_f = open(state, 'r+')
-        lock(state_f)
-
-        if args.all:
-            last_log_id = 0
-        else:
+            state_f = open(state, 'r+')
+            lock(state_f)
             last_log_id = int(state_f.readline())
-    else:
-        state_f = None
-        last_log_id = 0
+        else:
+            state_f = None
+            last_log_id = 0
 
-    with app.app_context():
-        last_log_id = export_patients(last_log_id)
+        with app.app_context():
+            if args.all:
+                last_log_id = export_all(last_log_id)
+            else:
+                last_log_id = export_changed(last_log_id)
 
-    if state_f is not None:
-        state_f.seek(0)
-        state_f.write(str(last_log_id))
-        state_f.write('\n')
-        state_f.truncate()
-        state_f.close()
+        if state_f is not None:
+            state_f.seek(0)
+            state_f.write(str(last_log_id))
+            state_f.write('\n')
+            state_f.truncate()
+            state_f.close()
+    elif args.patient_ids:
+        export_patients(args.patient_ids)
 
 
 if __name__ == '__main__':
